requirements.txt
streamlit
google-generativeai
faiss-cpu
langchain
PyPDF2
langchain_google_genai
python-dotenv
pypdf